pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Raza UL Mustafa", email = "mirxa999@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
studybuddy = "studybuddy:main"
run_agent = "studybuddy.studybuddy:run_agent"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
